Feature/extensions

+++ app/delete/delete.go
@@ -13,7 +13,15 @@ func Command() *cli.Command {
 		Action: func(c *cli.Context) error {
 			provider, cluster := app.MustCluster(c)
 
-			return provider.Delete(c.Context, cluster)
+			if ok := cli.MustConfirm("Are you sure you want to delete the cluster?", false); !ok {
+				return nil
+			}
+
+			cli.MustRun("Deleting Kubernetes Cluster...", func() error {
+				return provider.Delete(c.Context, cluster)
+			})
+
+			return nil
 		},
 	}
 }
